% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_politicos.R
\docType{data}
\name{data_politicos}
\alias{data_politicos}
\title{Data of every user in politicos en twitter shiny app (https://oderedes.shinyapps.io/politicosentwitter/)}
\format{
A dataset of 510 rows and 8 columns
\describe{
\item{user_id}{Character variable with the unique ID of the Twitter user}
\item{screen_name}{Character variable with the unique screen name of the Twitter user during may 2021}
\item{category}{Character variable that shows the category of the user into the national politics during may 2021}
\item{description}{Character variable that shows the description of the user in their account on Twitter during may 2021}
\item{name}{Character variable that shows the name of the user in their account on Twitter during may 2021}
\item{followers_count}{Numeric variable that shows the count of followers during may 2021}
\item{profile_image_url}{Character variable that shows the profile image of the user during may 2021}
\item{country}{Character variable that shows the country where the user were elected or born}
}
}
\source{
Politicos en Twitter Shiny App (https://oderedes.shinyapps.io/politicosentwitter/)
}
\usage{
data_politicos
}
\description{
A dataset containing information about every user downloaded
}
\keyword{datasets}
