% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_url.R
\name{download_url}
\alias{download_url}
\title{Get the url correspondent}
\usage{
download_url(x)
}
\arguments{
\item{x}{A number that describe the number of url path necessary}
}
\description{
With the right url we could access to the dataset storage
}
