% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends_followers.R
\name{get_friends_followers}
\alias{get_friends_followers}
\title{Download the friends and followers in a period of time}
\usage{
get_friends_followers(screen.name)
}
\arguments{
\item{screen.name}{A character vector with an Twitter user}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item date
\item followers_count
\item friends_count
\item listed_count
\item statuses_count
\item favourites_count
\item account_created_at
}
}
\description{
This download a time series with the evolution of friends, followers, listed, statuses, likes
and the created date of the account
}
\examples{
get_friends_followers(screen.name = "SergioMassa")
get_friends_followers(screen.name = c("SergioMassa", "dalesmm"))

}
\seealso{
\link[rtweet]{get_followers}

\link[rtweet]{get_friends}
}
