% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends_followers_all.R
\name{get_friends_followers_all}
\alias{get_friends_followers_all}
\title{Download the friends and followers in a period of time by category}
\usage{
get_friends_followers_all(category = "all", historic = T)
}
\arguments{
\item{category}{A character with the category selected -"deputies","senators","national executive","others","province servants", 'candidates',  "all"-}

\item{historic}{A logical value. Would you prefer the historic data?}

\item{country}{A character with the Alpha 2 code of the country that wants to download}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item date
\item followers_count
\item friends_count
\item listed_count
\item statuses_count
\item favourites_count
\item account_created_at
}
}
\description{
This download a time series with the evolution of friends, followers, listed, statuses, likes
and the created date of the account with the possibility of download the historic data by category
}
\examples{
get_friends_followers_all(category = 'others',  historic = TRUE)

}
\seealso{
get_friends_followers
}
