% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politicxsentwitter.R
\docType{package}
\name{politicxsentwitteR}
\alias{politicxsentwitteR}
\alias{politicxsentwitteR-package}
\title{politicxsentwitteR: Data of Politicians of Argentina on Twitter}
\description{
Back up and reusable data of what the principal public servants talk on Twitter
}
\author{
\strong{Maintainer}: Guadalupe Andrea Gonzalez \email{guadagonz12@gmail.com}

}
